requirements.txt
pandas
numpy
scipy
scikit-learn
joblib
pymongo
fastapi
uvicorn
prophet
xgboost
lightgbm
statsmodels
requests
python-dotenv
selenium
beautifulsoup4
